Edited the browse layout according to feedback and moved some files to new folders and locations.
