+++ react/src/components/CVToolbar.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { Button } from 'reactstrap'
 
 
-class EditButtonGroup extends Component {
+class CVToolbar extends Component {
 
   render() {
     return (
@@ -25,4 +25,4 @@ class EditButtonGroup extends Component {
 }
 
 
-export default EditButtonGroup
+export default CVToolbar
