--- react/src/components/Editor/EditorApp.css
@@ -1,5 +1,0 @@
-#textfields {
-  float:left;
-  width: 50%;
-  height: auto;
-}
