+++ react/src/components/NavBar.js
@@ -16,21 +16,4 @@ function NavBar() {
   )
 }
 
-/* class NavBar extends Component {
-  render() {
-    return (
-      <div className="navigation">
-        <Nav pills>
-          <NavItem>
-            <NavLink href="#">My CVs</NavLink>
-          </NavItem>
-          <NavItem>
-            <NavLink href="#">Browse</NavLink>
-          </NavItem>
-        </Nav>
-      </div>
-    )
-  }
-} */
-
 export default NavBar
