+++ react/src/components/Preview.js
