+++ react/src/components/SearchField.js
@@ -1,10 +1,8 @@
 import React, { Component } from 'react'
-
 class SearchField extends Component {
   handleChange(event) {
     this.props.updateUID(event.target.value)
   }
-
   render() {
     return (
       <div>
@@ -14,5 +12,4 @@ class SearchField extends Component {
     )
   }
 }
-
 export default SearchField
